# api/openapi.snapshot.json
{
  "components": {
    "schemas": {
...
      "GeocodeSuggestion": {
        "properties": {
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "lat": {
            "description": "Latitude (cast from LocationIQ string)",
            "title": "Lat",
            "type": "number"
          },
          "lon": {
            "description": "Longitude (cast from LocationIQ string)",
            "title": "Lon",
            "type": "number"
          }
        },
        "required": [
          "display_name",
          "lat",
          "lon"
        ],
        "title": "GeocodeSuggestion",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "JSON wrapper around the DFW air quality pipeline (PurpleAir + OpenAQ + IDW + traffic/wind adjustment).",
    "title": "AERIA \u00b7 DFW Air Quality API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/geocode/suggest": {
      "get": {
        "operationId": "suggest_api_geocode_suggest_get",
        "parameters": [
          {
            "description": "Partial address text",
            "in": "query",
            "name": "q",
            "required": true,
            "schema": {
              "description": "Partial address text",
              "minLength": 2,
              "title": "Q",
              "type": "string"
            }
          },
          {
            "description": "Max suggestions",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 5,
              "description": "Max suggestions",
              "maximum": 10,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/GeocodeSuggestion"
                  },
                  "title": "Response Suggest Api Geocode Suggest Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Suggest",
        "tags": [
          "geocode"
        ]
      }
    },
...
  }
}